requirements.txt
anyio==3.4.0
APScheduler==3.6.3
asgiref==3.4.1
cachetools==4.2.2
certifi==2021.10.8
click==8.0.3
fastapi==0.70.1
h11==0.12.0
idna==3.3
pydantic==1.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
python-telegram-bot==13.9
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
supervisor==4.2.2
tornado==6.1
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
uvicorn==0.16.0
